+++ eng/packages/http-client-csharp-mgmt/emitter/src/resource-detection.ts
@@ -56,7 +56,7 @@ export async function updateClients(
   );
   const resourceModels = getAllResourceModels(codeModel);
 
-  const resourceModelMap = new Map<string, ResourceMetadata>(
+  const resourceModelToMetadataMap = new Map<string, ResourceMetadata>(
     resourceModels.map((m) => [
       m.crossLanguageDefinitionId,
       {
@@ -67,10 +67,7 @@ export async function updateClients(
         ),
         resourceScope: getResourceScope(m),
         methods: [],
-        parentResource: getParentResourceModelId(
-          sdkContext,
-          models.get(m.crossLanguageDefinitionId)
-        )
+        parentResourceId: undefined // this will be populated later
       } as ResourceMetadata
     ])
   );
@@ -80,7 +77,7 @@ export async function updateClients(
 
   // then we iterate over all the clients and their methods to find the resource operations
   // and add them to the resource model metadata
-  // we also calculate the resource type from the path of the operation
+  // during the process we populate the resourceIdPattern and resourceType
   for (const client of clients) {
     for (const method of client.methods) {
       const serviceMethod = serviceMethods.get(
@@ -89,7 +86,7 @@ export async function updateClients(
       const [kind, modelId] =
         parseResourceOperation(serviceMethod, sdkContext) ?? [];
       if (modelId && kind) {
-        const entry = resourceModelMap.get(modelId);
+        const entry = resourceModelToMetadataMap.get(modelId);
         entry?.methods.push({
           id: method.crossLanguageDefinitionId,
           kind
@@ -106,9 +103,17 @@ export async function updateClients(
     }
   }
 
+  // after the resourceIdPattern has been populated, we can set the parentResourceId
+  for (const [modelId, metadata] of resourceModelToMetadataMap) {
+    const parentResourceModelId = getParentResourceModelId(sdkContext, models.get(modelId));
+    if (parentResourceModelId) {
+      metadata.parentResourceId = resourceModelToMetadataMap.get(parentResourceModelId)?.resourceIdPattern;
+    }
+  }
+
   // the last step, add the decorator to the resource model
   for (const model of resourceModels) {
-    const metadata = resourceModelMap.get(model.crossLanguageDefinitionId);
+    const metadata = resourceModelToMetadataMap.get(model.crossLanguageDefinitionId);
     if (metadata) {
       addResourceMetadata(sdkContext, model, metadata);
     }
@@ -281,7 +286,7 @@ function addResourceMetadata(
       code: "general-warning", // TODO -- later maybe we could define a specific code for resource hierarchy issues
       messageId: "default",
       format: {
-        message: `Cannot figure out resourceIdPatternResource from model ${model.name}.`
+        message: `Cannot figure out resourceIdPattern from model ${model.name}.`
       },
       target: NoTarget // TODO -- we need a method to find the raw target from the crossLanguageDefinitionId of this model
     });