+++ eng/packages/http-client-csharp-mgmt/generator/Azure.Generator.Management/src/Models/ResourceMetadata.cs
@@ -10,6 +10,6 @@ namespace Azure.Generator.Management.Models
         string ResourceType,
         ResourceScope ResourceScope,
         IReadOnlyList<ResourceMethod> Methods,
-        string? singletonResourceName,
-        string? ParentResource);
+        string? SingletonResourceName,
+        string? ParentResourceId);
 }