+++ eng/packages/http-client-csharp/generator/Azure.Generator/test/common/ModelTests.cs
@@ -73,17 +73,18 @@ namespace Azure.Generator.Tests.Common
             if (AssertFailures(strategy, format, serviceResponse, options))
                 return;
 
-            T model = (T)strategy.Read(serviceResponse, ModelInstance, options);
-
-            VerifyModel(model, format);
-            var data = strategy.Write(model, options);
+            var model = (T?)strategy.Read(serviceResponse, ModelInstance, options);
+            Assert.NotNull(model);
+            VerifyModel(model!, format);
+            var data = strategy.Write(model!, options);
             string roundTrip = data.ToString();
 
             // we validate those are equivalent element, representing the same json object (ignoring the spaces and orders, etc)
             AssertJsonEquivalency(expectedSerializedString, roundTrip);
 
-            T model2 = (T)strategy.Read(roundTrip, ModelInstance, options);
-            CompareModels(model, model2, format);
+            var model2 = (T?)strategy.Read(roundTrip, ModelInstance, options);
+            Assert.NotNull(model2);
+            CompareModels(model!, model2!, format);
         }
 
         private void AssertJsonEquivalency(string expected, string result)
@@ -167,12 +168,12 @@ namespace Azure.Generator.Tests.Common
                     if (strategy.GetType().Name.StartsWith("ModelJsonConverterStrategy"))
                     {
                         //we never get to the interface implementation because JsonSerializer errors before that
-                        Assert.Throws<JsonException>(() => { T model = (T)strategy.Read(serviceResponse, ModelInstance, options); });
+                        Assert.Throws<JsonException>(() => { var model = (T?)strategy.Read(serviceResponse, ModelInstance, options); });
                         result = true;
                     }
                     else
                     {
-                        Assert.Throws<InvalidOperationException>(() => { T model = (T)strategy.Read(serviceResponse, ModelInstance, options); });
+                        Assert.Throws<InvalidOperationException>(() => { var model = (T?)strategy.Read(serviceResponse, ModelInstance, options); });
                         result = true;
                     }
                 }
@@ -185,7 +186,7 @@ namespace Azure.Generator.Tests.Common
             }
             else if (ModelInstance is not IJsonModel<T> && format == "J")
             {
-                Assert.Throws<FormatException>(() => { T model = (T)strategy.Read(serviceResponse, ModelInstance, options); });
+                Assert.Throws<FormatException>(() => { T model = (T)strategy.Read(serviceResponse, ModelInstance, options)!; });
                 Assert.Throws<FormatException>(() => { var data = strategy.Write(ModelInstance, options); });
                 result = true;
             }