+++ sdk/communication/Azure.Communication.CallAutomation/src/Models/AnswerCallOptions.cs
@@ -56,5 +56,8 @@ namespace Azure.Communication.CallAutomation
         /// The operation context.
         /// </summary>
         public string OperationContext { get; set; }
+
+        /// <summary> Enables loopback audio functionality for the call. </summary>
+        public bool? EnableLoopbackAudio { get; set; }
     }
 }