+++ sdk/communication/Azure.Communication.CallAutomation/src/Models/SummarizationOptions.cs
@@ -1,0 +1,23 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Azure.Communication.CallAutomation
+{
+    /// <summary> Configuration options for call summarization. </summary>
+    public class SummarizationOptions
+    {
+        /// <summary> Indicating whether end call summary should be enabled. </summary>
+        public bool? EnableEndCallSummary { get; set; }
+        /// <summary> Locale for summarization (e.g., en-US). </summary>
+        internal string Locale { get; set; }
+
+        /// <summary>
+        /// Locale for summarization (e.g., en-US).
+        /// </summary>
+        /// <param name="locale"></param>
+        public SummarizationOptions(string locale)
+        {
+            Locale = locale;
+        }
+    }
+}