+++ sdk/communication/Azure.Communication.CallAutomation/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release History
 
+## 1.6.0-beta.1 (Unreleased)
+
+### Features Added
+- Added support for move participant in a call to another call.
+- Added support for Get recording result.
+- Added support for Teams phone call details with custom calling context.
+- Added support for PiiRedaction and summarization in transcription.
+- Added support for sentiments analysis in speech and choice recognition.
+
 ## 1.5.0 (Unreleased)
 
 ### Features Added