+++ sdk/storage/Microsoft.Azure.WebJobs.Extensions.Storage.Blobs/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Breaking Changes
 
 ### Bugs Fixed
+- Fixed bug where the scan will continue scanning AzureWebJobsStorage even when configuring a target storage account
 
 ### Other Changes